Gjør ISanityBegrunnelse sealed
