+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/domene/eøs/EØSBegrunnelseMedTriggere.kt
@@ -1,7 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.brev.domene.eøs
 
+import no.nav.familie.ba.sak.kjerne.brev.domene.SanityEØSBegrunnelse
 import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.EØSStandardbegrunnelse
-import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.SanityEØSBegrunnelse
 
 data class EØSBegrunnelseMedTriggere(
     val eøsBegrunnelse: EØSStandardbegrunnelse,
