+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/begrunnelser/SanityEØSBegrunnelse.kt
@@ -1,10 +1,8 @@
 package no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser
 
-import no.nav.familie.ba.sak.kjerne.brev.domene.ISanityBegrunnelse
+import no.nav.familie.ba.sak.kjerne.brev.domene.SanityEØSBegrunnelse
 import no.nav.familie.ba.sak.kjerne.brev.domene.SanityPeriodeResultat
 import no.nav.familie.ba.sak.kjerne.brev.domene.Tema
-import no.nav.familie.ba.sak.kjerne.brev.domene.UtvidetBarnetrygdTrigger
-import no.nav.familie.ba.sak.kjerne.brev.domene.VilkårTrigger
 import no.nav.familie.ba.sak.kjerne.brev.domene.finnEnumverdi
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.KompetanseAktivitet
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.KompetanseResultat
@@ -71,28 +69,4 @@ data class RestSanityEØSBegrunnelse(
 
     private inline fun <reified T> konverterTilEnumverdi(it: String): T? where T : Enum<T> =
         enumValues<T>().find { enum -> enum.name == it }
-}
-
-data class SanityEØSBegrunnelse(
-    override val apiNavn: String,
-    override val navnISystem: String,
-    override val periodeResultat: SanityPeriodeResultat? = null,
-    override val vilkår: Set<Vilkår>,
-    override val fagsakType: FagsakType?,
-    override val tema: Tema?,
-    val annenForeldersAktivitet: List<KompetanseAktivitet>,
-    val barnetsBostedsland: List<BarnetsBostedsland>,
-    val kompetanseResultat: List<KompetanseResultat>,
-    val hjemler: List<String>,
-    val hjemlerFolketrygdloven: List<String>,
-    val hjemlerEØSForordningen883: List<String>,
-    val hjemlerEØSForordningen987: List<String>,
-    val hjemlerSeperasjonsavtalenStorbritannina: List<String>,
-) : ISanityBegrunnelse {
-    override val lovligOppholdTriggere: List<VilkårTrigger> = emptyList()
-    override val utvidetBarnetrygdTriggere: List<UtvidetBarnetrygdTrigger> = emptyList()
-    override val valgbarhet = null
-    override val bosattIRiketTriggere: List<VilkårTrigger> = emptyList()
-    override val giftPartnerskapTriggere: List<VilkårTrigger> = emptyList()
-    override val borMedSokerTriggere: List<VilkårTrigger> = emptyList()
 }
