+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/config/SanityKlientMock.kt
@@ -3,9 +3,9 @@ package no.nav.familie.ba.sak.config
 import no.nav.familie.ba.sak.integrasjoner.sanity.SanityKlient
 import no.nav.familie.ba.sak.kjerne.brev.domene.ISanityBegrunnelse
 import no.nav.familie.ba.sak.kjerne.brev.domene.SanityBegrunnelse
+import no.nav.familie.ba.sak.kjerne.brev.domene.SanityEØSBegrunnelse
 import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.EØSStandardbegrunnelse
 import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.IVedtakBegrunnelse
-import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.SanityEØSBegrunnelse
 import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.Standardbegrunnelse
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.context.annotation.Bean
